+++ election/models.py
@@ -1,7 +1,9 @@
-from django.db import models, IntegrityError
-from django.contrib.postgres.fields import ArrayField
-from libs.django_randomprimary import RandomPrimaryIdModel
 import logging
+
+from django.contrib.postgres.fields import ArrayField
+from django.db import IntegrityError, models
+
+from libs.django_randomprimary import RandomPrimaryIdModel
 
 logger = logging.getLogger(__name__)
 
@@ -12,11 +14,17 @@ class Election(RandomPrimaryIdModel):
     title = models.CharField("Title", max_length=255)
     candidates = ArrayField(models.CharField("Name", max_length=255))
     on_invitation_only = models.BooleanField(default=False)
-    num_grades = models.SmallIntegerField("Num. grades")
+    num_grades = models.PositiveSmallIntegerField("Num. grades", null=False)
     
     # make sure we don't ask for more grades than allowed in the database
     def save(self, *args, **kwargs):
 
+        if self.num_grades is None:
+            raise IntegrityError("Election requires a positive number of grades.")
+        
+        if self.title is None or self.title == "":
+            raise IntegrityError("Election requires a proper title")
+            
         if self.num_grades > MAX_NUM_GRADES or self.num_grades <= 0:
             raise IntegrityError(
                 "Max number of grades is %d. Asked for %d grades"
